# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Interactive API for testing Neuron keylib functionality",
        "title": "Keylib API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/mnemonic/derive-key": {
            "post": {
                "description": "Derive a private key from a BIP39 mnemonic using default path (m/44'/60'/0'/0/0)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mnemonic"
                ],
                "summary": "Derive key from mnemonic",
                "parameters": [
                    {
                        "description": "Mnemonic phrase",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.DeriveFromMnemonicRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.KeyPairResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/mnemonic/derive-key-full": {
            "post": {
                "description": "Derive a private key from a BIP39 mnemonic with custom passphrase and derivation path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mnemonic"
                ],
                "summary": "Derive key with options",
                "parameters": [
                    {
                        "description": "Mnemonic with options",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.DeriveFromMnemonicFullRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.KeyPairResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/mnemonic/generate": {
            "post": {
                "description": "Generate a new BIP39 mnemonic phrase with specified word count",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mnemonic"
                ],
                "summary": "Generate a new mnemonic",
                "parameters": [
                    {
                        "description": "Word count (12, 15, 18, 21, or 24)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.GenerateMnemonicRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.MnemonicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/mnemonic/validate": {
            "post": {
                "description": "Validate a BIP39 mnemonic phrase",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mnemonic"
                ],
                "summary": "Validate a mnemonic",
                "parameters": [
                    {
                        "description": "Mnemonic phrase",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ValidateMnemonicRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ValidateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.DeriveFromMnemonicFullRequest": {
            "type": "object",
            "required": [
                "mnemonic"
            ],
            "properties": {
                "mnemonic": {
                    "type": "string",
                    "example": "abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon about"
                },
                "passphrase": {
                    "type": "string",
                    "example": ""
                },
                "path": {
                    "type": "string",
                    "example": "m/44'/60'/0'/0/0"
                }
            }
        },
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.DeriveFromMnemonicRequest": {
            "type": "object",
            "required": [
                "mnemonic"
            ],
            "properties": {
                "mnemonic": {
                    "type": "string",
                    "example": "abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon about"
                }
            }
        },
...
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid hex format"
                }
            }
        },
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.GenerateMnemonicRequest": {
            "type": "object",
            "required": [
                "wordCount"
            ],
            "properties": {
                "wordCount": {
                    "type": "integer",
                    "enum": [
                        12,
                        15,
                        18,
                        21,
                        24
                    ],
                    "example": 12
                }
            }
        },
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.KeyPairResponse": {
            "type": "object",
            "properties": {
                "evmAddress": {
                    "type": "string",
                    "example": "0xe364f2f1e5f4f03d1df682322500b9c68c997ec3"
                },
                "evmAddressChecksum": {
                    "type": "string",
                    "example": "0xe364f2f1e5F4F03d1df682322500b9c68C997ec3"
                },
                "peerID": {
                    "type": "string",
                    "example": "16Uiu2HAm3Lkn9NRieuh3UUTWMNthSDumQL9ctTBKxQqdCC79WUSq"
                },
                "privateKeyHex": {
                    "type": "string",
                    "example": "0x1234567890abcdef..."
                },
                "publicKeyHex": {
                    "type": "string",
                    "example": "0x02759b048e7ccf6ba68f9658105a4a139b5f9f5dfd451857c600cc28f33a1a99ae"
                },
                "publicKeyHexUncompressed": {
                    "type": "string",
                    "example": "0x04759b048e7ccf6ba68f9658105a4a139b5f9f5dfd451857c600cc28f33a1a99ae..."
                }
            }
        },
...
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.MnemonicResponse": {
            "type": "object",
            "properties": {
                "mnemonic": {
                    "type": "string",
                    "example": "abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon about"
                },
                "wordCount": {
                    "type": "integer",
                    "example": 12
                }
            }
        },
...
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ValidateMnemonicRequest": {
            "type": "object",
            "required": [
                "mnemonic"
            ],
            "properties": {
                "mnemonic": {
                    "type": "string",
                    "example": "abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon about"
                }
            }
        },
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ValidateResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": ""
                },
                "valid": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    }
}